requirements.txt
python-docx
pdf2image
pytesseract
opencv-python
joblib
scikit-learn
pandas
matplotlib
streamlit
